Add "ParsedDocumentData" and do not rely on anonymous class in "Parser"
